\name{tensorMam-package}
\alias{tensorMam-package}
\alias{tensorMam}
\docType{package}
\title{
  A tensor estimation approach to multivariate additive models
}
\description{
  For a high-dimensional multivariate additive model (MAM) using B-splines, with or without aparsity assumptions, 
  treating the coefficients as a third-order tensor and borrowing Tucker decomposition to reduce the number of parameters.  
  The multivariate sparse group lasso (mcp or scad) and the coordinate descent algorithm are used to estimate
  functions for sparsity situation.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Xu Liu

Maintainer: Xu Liu <liu.xu@sufe.edu.cn>
}
\references{
  A tensor estimation approach to multivariate additive models.
}
\keyword{ High-dimensional, Sparse models; Tensor estimation; Tucker decomposition. }
